Move process package integration tests to own files
