+++ executors/kubernetes/util_test.go
@@ -148,6 +148,7 @@ func TestGetKubeClientConfig(t *testing.T) {
 			},
 		},
 	}
+
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
 			inClusterConfig = test.inClusterConfig
