+++ helpers/process/killer_windows_integration_test.go
@@ -1,4 +1,6 @@
-package process
+// +build windows
+
+package process_test
 
 // Cases for Windows that are used in `filler.go#TestKiller`.
 func testKillerTestCases() map[string]testKillerTestCase {
